게임 구조 재설정
